+++ dummy_client/attach_dummy.py
@@ -4,7 +4,7 @@ sys.path.insert(0, "/home/bigs/battle.ai/")
 print sys.path
 
 from multi_dummy import DummyManager
-from gameLogic.pixels.MyPixelsParser2 import MyPixelsParser2
+from game.pixels import MyPixelsParser2
 
 dm = DummyManager(MyPixelsParser2, 'Dummy3')
 dm.attach('104.199.218.103', 9001)
