+++ game/baskin/baskinServer.py
@@ -1,7 +1,7 @@
 import sys
 sys.path.insert(0, '../')
-from gameLogic.baseClass.TurnGameLogic import TurnGameLogic
-from gameLogic.baseClass.Phase import Phase
+from gamebase.game.TurnGameLogic import TurnGameLogic
+from gamebase.game.Phase import Phase
 import logging
 logging.basicConfig(level=logging.DEBUG)
