+++ game/othello/OthelloProtocol
