--- gameLogic/st.py
@@ -1,3 +1,0 @@
-
-
-print '안녕'
